requirements.txt
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
asn1crypto==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.17.0
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
flake8==4.0.1
flake8-docstrings==1.6.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
genmsg==0.5.16
genpy==0.6.16
gpg==1.16.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifcfg==0.18
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
keyring==23.5.0
keyrings.alt==4.1.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mypy==0.942
mypy-extensions==0.4.3
netifaces==0.11.0
nose==1.3.7
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauth==1.0.1
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
patiencediff==0.2.1
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
pytest-timeout==2.1.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
roman==3.3
rosclean==1.15.8
rosdep==0.24.0
rosdep-modules==0.24.0
rosdistro-modules==0.9.1
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
rosmaster==1.15.14
rosparam==1.15.14
rospkg==1.5.1
rospkg-modules==1.5.1
scipy==1.8.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sympy==1.9
system-service==0.3
systemd-python==234
toml==0.10.2
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vcstool==0.3.0
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
